handle roles and permissions for users and policies for assessment and result and gates for foundation and participant

+++ app/Http/Controllers/AssessmentController.php
@@ -6,8 +6,10 @@ use App\DataTables\FoundationResultsDataTable;
 use App\Http\Requests\StoreAssessmentRequest;
 use App\Http\Requests\UpdateAssessmentRequest;
 use App\Models\Assessment;
+use App\Models\Result;
 use App\Services\AssessmentService;
 use App\Services\CategoryService;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\View;
 
 class AssessmentController extends Controller
@@ -26,6 +28,8 @@ class AssessmentController extends Controller
      */
     public function create()
     {
+        Gate::authorize('create', Assessment::class);
+
         $categories = app(CategoryService::class)->index();
 
         return view('pages.assessment.create', compact('categories'));
@@ -36,6 +40,8 @@ class AssessmentController extends Controller
      */
     public function store(StoreAssessmentRequest $request)
     {
+        Gate::authorize('create', Assessment::class);
+
         $this->assessmentService->store($request->validated());
 
         return redirect()->route('profile.index')->with('success', 'Assemmentt Created Successfully.');
@@ -46,11 +52,12 @@ class AssessmentController extends Controller
      */
     public function show(Assessment $assessment, FoundationResultsDataTable $foundationResultsDataTable)
     {
+        Gate::authorize('view', Assessment::class);
+        
         $assessment->load('questions', 'results');
         $categories = app(CategoryService::class)->index();
         
         return $foundationResultsDataTable->render('pages.assessment.show', compact('assessment', 'categories'));
-        // return view('pages.assessment.show', compact('assessment'));
     }
 
     /**
@@ -66,6 +73,8 @@ class AssessmentController extends Controller
      */
     public function update(UpdateAssessmentRequest $request, Assessment $assessment)
     {
+        Gate::authorize('update', Assessment::class);
+
         $this->assessmentService->update($assessment, $request->validated());
 
         return redirect()->route('assessments.show', $assessment->id)->with('success', 'assessment updated successfully.');
@@ -76,6 +85,8 @@ class AssessmentController extends Controller
      */
     public function destroy(Assessment $assessment)
     {
+        Gate::authorize('delete', Assessment::class);
+
         $this->assessmentService->destroy($assessment);
 
         return redirect()->back()->with('success', 'Assessment Deleted Successfully.');
@@ -83,6 +94,8 @@ class AssessmentController extends Controller
 
     public function policy(Assessment $assessment)
     {
+        Gate::authorize('create', Result::class);
+
         return view('pages.assessment.policy', compact('assessment'));
     }
 }
